requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic-settings==2.7.0
python-multipart==0.0.19
aiofiles==24.1.0
langchain==0.3.14
langchain-openai==0.3.1
langfuse>=2.0.0
langchain-community==0.3.14
chromadb==0.5.23
httpx==0.28.1
pypdf2==3.0.1
pytest==8.3.4
pytest-asyncio==0.25.0
easyocr==1.7.2
Pillow==11.1.0
